chore: another refactor for another circular dependency
